requirements.txt
transformers
torch
pymupdf
python-docx
